more tests

+++ get_next_line.h
@@ -27,7 +27,7 @@
 int get_next_line(int fd, char **line);
 
 size_t  ft_strlen(const char *str);
-void	*ft_memcpy(void *dst, const void *src, size_t n);
+void	*ft_memcpy(void *dst, const void *src, int n);
 char	*ft_strdup(const char *s1);
 char	*ft_strjoin(char const *s1, char const *s2);
 char	*ft_substr(char const *s, unsigned int start, size_t len);
