+++ get_next_line_utils.c
@@ -22,9 +22,9 @@ size_t		ft_strlen(const char *str)
 	return (i);
 }
 
-void	*ft_memcpy(void *dst, const void *src, size_t n)
+void	*ft_memcpy(void *dst, const void *src, int n)
 {
-	size_t				i;
+	int				i;
 	unsigned char		*d;
 	const unsigned char *s;
 
@@ -95,17 +95,6 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	sub[i] = '\0';
 	return (sub);
 }
-
-
-
-
-
-
-
-
-
-
-
